:sparkles: #202 Feat: 멤버 객체 비교 로직 수정
